requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7
SQLAlchemy==2.0.21
sentence-transformers==2.7.0
faiss-cpu==1.7.4
openai==0.28.1
google-generativeai==0.8.3
PyMuPDF==1.23.8
python-docx==0.8.11
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
reportlab==4.4.3
